cleans up modal implementatino
