+++ src/main.js
@@ -4,8 +4,8 @@ import "@graph-paper/core/style.css";
 const app = new App({
   target: document.body,
   props: {
-    name: "Growth & Usage Dashboard"
-  }
+    name: "Growth & Usage Dashboard",
+  },
 });
 
 export default app;
